Add color, some text styling, & assets to page
The wireframe was mostly done (just need to fix the bug with the page
exceeding 100vh), so the "paint" was added so to speak :) Color and
shadows I prioritized since it'll add the most depth to the page given
the level of effort. Next will prioritize making the text content look
nice.